+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Store degrees in LocalStorage
+
 ### Removed
 
 ## [1.9.2] - 08/09/2021